refactor: split CLIcore_UI_builtins.c and CLIcore_UI_completion.c under 1000 lines
Phase 2 of file-size refactoring to enforce <1000 lines per C file.

CLIcore_UI_builtins.c (1021 lines) split into:
  - CLIcore_UI_builtins.c (774 lines): core builtins
  - CLIcore_UI_builtins_fifo.c (283 lines): FIFO mgmt + list cmds

CLIcore_UI_completion.c (1034 lines) split into:
  - CLIcore_UI_completion.c (361 lines): prompt, input, levenshtein
  - CLIcore_UI_completion_tab.c (838 lines): tab-completion engine
