+++ src/cli/CLIcore/CMakeLists.txt
@@ -30,7 +30,9 @@ set(CORE_FILES
     CLIcore/CLIcore_help_tui.c
     CLIcore/CLIcore_UI_bookmark.c
     CLIcore/CLIcore_UI_builtins.c
+    CLIcore/CLIcore_UI_builtins_fifo.c
     CLIcore/CLIcore_UI_completion.c
+    CLIcore/CLIcore_UI_completion_tab.c
     CLIcore/CLIcore_UI_highlight.c
     CLIcore/CLIcore_UI_hintarea.c
     CLIcore/CLIcore_UI_history.c
